refactor: use toArray(arr)[0] to get a child
